% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAPCG.R
\name{plotAPCG}
\alias{plotAPCG}
\title{Plot All Protein Composition Graph}
\usage{
plotAPCG(file = "proSeq")
}
\arguments{
\item{file}{input csv file containing the list of pdb id and sequence}
}
\value{
This function returns a matrix of composition descriptor infomation as well as
a full composition graph given the proteins.
}
\description{
A function that plots a comprehensive composition mapping given a protein list.
The composition descriptors includes hydrophobicity, van der Waals volume,
polarity, polarizability and desolvation. A hierarchical row and column clustering
will also be displayed on the side to showing the similarity analysis
}
\examples{
composition_plot <- plotAPCG(file = "proSeq")
composition_plot

}
\references{
Xiao, N., Cao, D.-S., Zhu, M.-F., & Xu, Q.-S. (2015). protr/ProtrWeb: R package and web server for
generating various numerical representation schemes of protein sequences. Bioinformatics, 31(11), 1857–1859.
https://doi.org/10.1093/bioinformatics/btv042
}
