% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCG.R
\name{plotCG}
\alias{plotCG}
\title{Plot Composition Graph}
\usage{
plotCG(file = system.file("data/proSeq.rda", package = "Rppsc"),
  type = 1)
}
\arguments{
\item{file}{input csv file containing the protein name and sequence}

\item{type}{integer, indicates one attribute for plotting the composition
different Attributes are listed below:
1: hydrophobicity
2: van der Waals volume
3: polarity
4: polarizability
5: desolvation}
}
\value{
This function returns a composition graph given the attribute type.
}
\description{
A function that plots the composition of protein sequence given one attribute
at a time. Attributes include hydrophobicity, van der Waals volume, polarity,
polarizability, and desolvation
}
\examples{
f = system.file("data/proSeq.rda",package = "Rppsc")
p <- function(f,type = 1)
print(p)

}
