% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proSeq}
\alias{proSeq}
\title{Protein Sequence}
\format{A table format:
\describe{
 \tabular{ll}{
  pdb id  \tab sequence \cr
  3blr \tab IGQGTFGEVFKARHRIKIDFLGKTKEGFPITALREIKILQLLKH \cr
  2wtc \tab LGSEVKLAFERKTCKKVAIKILKKLNHPCIIKIKNFFD \cr
  4qsv \tab VFTKPVDPDEVPDYVTVIDLSLHK \cr
  5ndd \tab LPIVYTIVFVVALPVFLFRTKKKAPAVIYMANLALAD
  }
}}
\source{
Structural Genomics Consortium lab, Canada.
}
\usage{
proSeq
}
\description{
This dataset include a list of protein id(pdb code) and sequence.
}
\examples{
\dontrun{
 data(proSeq)
}
}
\keyword{datasets}
