% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRppsc.R
\name{runRppsc}
\alias{runRppsc}
\title{launch Rppsc shiny app}
\usage{
runRppsc()
}
\value{
null
}
\description{
A function launches the Rppsc Shiny app that allows the user to use the features
of Rppsc package in an interactive way.
}
\examples{
\dontrun{
  runRppsc()
}


}
