% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAPCG.R
\name{plotAPCG}
\alias{plotAPCG}
\title{Plot All Protein Composition Graph}
\usage{
plotAPCG(file = "protrdata/pdbSeq.csv")
}
\arguments{
\item{file}{input csv file containing list of protein name and sequence}
}
\value{
This function returns a full composition graph given the proteins.
}
\description{
A function that plots a comprehensive composition mapping given protein list.
The composition descriptors includes hydrophobicity, van der Waals volume,
polarity, polarizability and desolvation.
}
\examples{
p <- plotAPCG(file = "data/pdbSeq.csv")
print(p)

}
