% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCG.R
\name{plotCG}
\alias{plotCG}
\title{Plot A protein Composition Graph}
\usage{
plotCG(file = "proSeq", type = 1, circular_plot = TRUE)
}
\arguments{
\item{file}{input csv file containing the protein name and sequence}

\item{type}{integer, indicates one attribute for plotting the composition
different Attributes are listed below:
1: hydrophobicity
2: van der Waals volume
3: polarity
4: polarizability
5: desolvation}
}
\value{
This function returns a composition graph given the attribute type.
}
\description{
A function that plots the composition of protein sequence given one attribute
at a time. Attributes include hydrophobicity, van der Waals volume, polarity,
polarizability, and desolvation
}
\examples{
# defaul value is to show hydrophobicity composition in a circular plot
p <- plotCG(file = "proSeq",type= 1, circular_plot = TRUE )
p #call p to see the plot

}
\references{
The R Graph Gallery -Circular barplot with groups. (2018).Retrieved
December 1, 2019, from https://www.r-graph-gallery.com/297-circular-barplot-with-groups.html
}
